+++ spec/examples/pay_with_provider_advanced_no_message.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class PayWithProviderAdvancedNoMessage < Actor
+  input :provider,
+        inclusion: {
+          in: %w[MANGOPAY PayPal Stripe]
+        }
+end